+++ test/logger_test.js
@@ -10,11 +10,13 @@ var log = require('../src/logger');
 
 describe('logger', function() {
   it('should return the default parent logger if no config is available', function() {
+    log.init({});
     var logger = log.getLogger('myLogger');
     expect(logger).to.be.an.instanceOf(bunyan);
   });
 
-  it('should return a child logger if wanted', function() {
+  it('should return a child logger if requested', function() {
+    log.init({});
     var logger = log.getLogger('childName');
     expect(logger).to.be.an.instanceOf(bunyan);
     expect(logger.fields).to.have.property('module');
@@ -74,5 +76,40 @@ describe('logger', function() {
 
     var logger = log.getLogger('myLogger');
     expect(logger).not.to.be.an.instanceOf(bunyan);
+  });
+
+  it('should reset loggers when the logger is set after initialization', function() {
+    log.init({});
+    var reInitCalled = false;
+    var logger;
+    logger = log.getLogger('myLogger', function(newLogger) {
+      reInitCalled = true;
+      logger = newLogger;
+    });
+
+    // first getLogger call should yield the default bunyan logger
+    expect(logger).to.be.an.instanceOf(bunyan);
+    expect(logger.fields.name).to.equal('instana-nodejs-sensor');
+    var originalLogger = logger;
+
+    var logger2 = bunyan.createLogger({ name: 'new-logger' });
+    log.init({ logger: logger2 });
+
+    expect(reInitCalled).to.be.true;
+    expect(logger).to.be.an.instanceOf(bunyan);
+    expect(logger === originalLogger).to.not.be.true;
+    expect(logger.fields.name).to.equal('new-logger');
+  });
+
+  it('should not choke on re-initialization when there is no reInit callback', function() {
+    log.init({});
+    var logger = log.getLogger('myLogger');
+
+    // first getLogger call should yield the default bunyan logger
+    expect(logger).to.be.an.instanceOf(bunyan);
+    expect(logger.fields.name).to.equal('instana-nodejs-sensor');
+
+    var logger2 = bunyan.createLogger({ name: 'new-logger' });
+    log.init({ logger: logger2 });
   });
 });