+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unpublished
+- Provide API to set a logger after initialization to resolve init/logger cycle.
+
 ## 1.56.0
 - Record [koa-router](https://github.com/alexmingoia/koa-router) routes as path templates when tracing HTTP entries.
 - Improve announce payload.